third_party/blink/web_tests/virtual/stable/external/wpt/css/css-animations/parsing/animation-range-shorthand-expected.txt

This is a testharness.js-based test.
Found 57 tests; 21 PASS, 36 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL e.style['animation-range'] = "cover" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "contain" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter, exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter 0% enter 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "exit 0% exit 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "cover 0% cover 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "contain 0% contain 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter calc(10% - 10%) enter calc(50% + 50%)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "cover 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "contain 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter 50px exit 100px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "exit calc(10% + 50px)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "enter 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-range'] = "cover 50% enter 50%, contain 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['animation-range'] = "enter 50% 0s" should not set the property value
PASS e.style['animation-range'] = "0s enter 50%" should not set the property value
PASS e.style['animation-range'] = "1s" should not set the property value
PASS e.style['animation-range'] = "-1s" should not set the property value
PASS e.style['animation-range'] = "1s 2s" should not set the property value
PASS e.style['animation-range'] = "1s, 2s" should not set the property value
PASS e.style['animation-range'] = "1s 2s, 3s" should not set the property value
PASS e.style['animation-range'] = "1s, 2s 3s" should not set the property value
PASS e.style['animation-range'] = "1s, 2s, 3s" should not set the property value
PASS e.style['animation-range'] = "1s 2s 3s" should not set the property value
PASS e.style['animation-range'] = "0s, 1s 2s 3s" should not set the property value
PASS e.style['animation-range'] = "1s / 2s" should not set the property value
PASS e.style['animation-range'] = "1s, 2px" should not set the property value
PASS e.style['animation-range'] = "#ff0000" should not set the property value
PASS e.style['animation-range'] = "red" should not set the property value
PASS e.style['animation-range'] = "thing" should not set the property value
PASS e.style['animation-range'] = "thing 0%" should not set the property value
PASS e.style['animation-range'] = "thing 42%" should not set the property value
PASS e.style['animation-range'] = "thing 100%" should not set the property value
PASS e.style['animation-range'] = "thing 100px" should not set the property value
PASS e.style['animation-range'] = "100% thing" should not set the property value
FAIL e.style['animation-range'] = "cover" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "cover 100%" but got (undefined) undefined
FAIL e.style['animation-range'] = "cover" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "cover 0%" but got (undefined) undefined
FAIL e.style['animation-range'] = "cover" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-range'] = "contain" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "contain 100%" but got (undefined) undefined
FAIL e.style['animation-range'] = "contain" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "contain 0%" but got (undefined) undefined
FAIL e.style['animation-range'] = "contain" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-range'] = "enter 10% exit 20%" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "exit 20%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter 10% exit 20%" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "enter 10%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter 10% exit 20%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-range'] = "enter calc(10% + 10px) exit 20px" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "exit 20px" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter calc(10% + 10px) exit 20px" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "enter calc(10% + 10px)" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter calc(10% + 10px) exit 20px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-range'] = "enter, exit" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "enter 100%, exit 100%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter, exit" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "enter 0%, exit 0%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter, exit" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-range'] = "enter 0%, exit" should set animation-range-end assert_equals: animation-range-end should be canonical expected (string) "auto, exit 100%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter 0%, exit" should set animation-range-start assert_equals: animation-range-start should be canonical expected (string) "enter 0%, exit 0%" but got (undefined) undefined
FAIL e.style['animation-range'] = "enter 0%, exit" should not set unrelated longhands assert_true: expected true got false
Harness: the test ran to completion.
